UI refactoring. Removed crud repositories. Added possibility for the organizer to remove a member. Added a notification system on backend and ui

+++ lib/model/data_models/juror_vote.dart
@@ -5,7 +5,7 @@ class JurorVote extends Equatable {
   final DateTime createdAt;
   final String jurorVotingId;
   final String votingFormFieldId;
-  final int value;
+  final double value;
 
   const JurorVote({
     required this.id,
@@ -21,7 +21,7 @@ class JurorVote extends Equatable {
       createdAt: DateTime.parse(json['created_at']).toLocal(),
       jurorVotingId: json['juror_voting_id'] as String,
       votingFormFieldId: json['voting_form_field_id'] as String,
-      value: json['value'] as int,
+      value: json['value'] as double,
     );
   }
 
@@ -50,7 +50,7 @@ class JurorVote extends Equatable {
     DateTime? createdAt,
     String? jurorVotingId,
     String? votingFormFieldId,
-    int? value,
+    double? value,
   }) {
     return JurorVote(
       id: id ?? this.id,
